% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat_rungano.R
\docType{data}
\name{dat_rungano}
\alias{dat_rungano}
\title{The dat_rungano data set}
\format{A data frame}
\usage{
data(dat_rungano)
}
\description{
\itemize{
\item{t1}{ something}
\item{t2}{something}
\item{t3}{something }
\item{t4}{something }
\item{y1}{ something}
\item{y2}{ something}
\item{y3}{ something}
\item{y4}{something}
\item{se1}{ something}
\item{se2}{ something}
\item{se3}{ something}
\item{se4}{ something}
\item{na}{tsomething}
\item{s}{tsomething}
\item{prsv}{tsomething}
\item{Author}{tsomething}
\item{rand.allo}{tsomething}
\item{blinding}{tsomething}
\item{peerRev}{tsomething}
\item{type}{tsomething}
}
}
\keyword{datasets}
