% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest_metafor.R
\name{forest_metafor}
\alias{forest_metafor}
\title{Forest plot from metafor package translate to ggplot2}
\usage{
forest_metafor(x, annotate = TRUE, addfit = TRUE, addcred = FALSE, 
showweights = FALSE, xlim, alim, clim, ylim, at, steps = 5, level = x$level,
 refline = 0, digits = 2L, width, xlab, slab, mlab, ilab, ilab.xpos, ilab.pos,
  order, transf, atransf, targs, rows, efac = 1, pch = 15,psize, col, border,
   lty, cex, cex.lab, cex.axis, ...)
}
\arguments{
\item{x}{da,d.a/,dsf.a}

\item{annotate}{aaaa}

\item{addfit}{aaa}

\item{addcred}{aaaa}

\item{showweights}{aaaa}

\item{xlim}{aaaa}

\item{alim}{aaaa}

\item{clim}{aaaa}

\item{ylim}{aaaa}

\item{at}{aaaa}

\item{steps}{aaaa}

\item{level}{aaaa}

\item{refline}{aaaa}

\item{digits}{aaaa}

\item{width}{aaaa}

\item{xlab}{aaaa}

\item{slab}{aaaa}

\item{mlab}{aaaa}

\item{ilab}{aaaa}

\item{ilab.xpos}{aaaa}

\item{ilab.pos}{aaaa}

\item{order}{aaaa}

\item{transf}{aaaa}

\item{atransf}{aaaa}

\item{targs}{aaaa}

\item{rows}{aaaa}

\item{efac}{aaaa}

\item{pch}{aaaa}

\item{psize}{aaaa}

\item{col}{aaaa}

\item{border}{aaaa}

\item{lty}{aaaa}

\item{cex}{aaaa}

\item{cex.lab}{aaaa}

\item{cex.axis}{aaaa}

\item{...}{optional argument to}
}
\value{
returns to
}
\description{
Forest plot from metafor package translate to ggplot2
}
