% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaforest_plot.R
\name{metaforest_plot}
\alias{metaforest_plot}
\title{Implements a forest plot using ggplot2 package.}
\usage{
metaforest_plot(data, Treatment, effect, cilb, ciub, size = 2)
}
\arguments{
\item{data}{Data frame with the information needed for the forest plot (effect, confidence interval and reference name).}

\item{Treatment}{column name  with the cite information or treatment name.}

\item{effect}{column name with the observed effect sizes or outcomes.}

\item{cilb}{column name with the lower bound ci.}

\item{ciub}{column name with the upper bound ci.}

\item{size}{points size}
}
\description{
Implements a forest plot using ggplot2 package.
}
