% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaupdate.R
\name{metaupdate}
\alias{metaupdate}
\title{Meta-analysis shiny app}
\usage{
metaupdate(datapair, pair_result,trt.pair)
}
\arguments{
\item{datapair}{Data frame with .....}

\item{pair_result}{list with the pairwise meta-analysis models}

\item{trt.pair}{variable with the pairwise treatment names}
}
\value{
shiny app.
}
\description{
\code{metaupdate} implements .....
}
\examples{
metaupdate(MTCpairs2, pair_result, trt.pair)
}

