% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modstr.R
\docType{data}
\name{modstr}
\alias{modstr}
\title{The data structure to run the app}
\format{list}
\usage{
data(modstr)
}
\description{
\itemize{
\item{list}{ something}
\item{list}{something}
}
}
\keyword{datasets}
