% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTCdata.R
\docType{data}
\name{MTCdata}
\alias{MTCdata}
\title{The MTCdata data set}
\format{A data frame}
\usage{
data(MTCdata)
}
\description{
\itemize{
\item{event1}{ something}
\item{event2}{something}
\item{event3}{something }
\item{n1}{something }
\item{n2}{ something}
\item{n3}{ something}
\item{treat1}{ something}
\item{treat2}{something}
\item{event2}{ something}
\item{treat3}{ something}
}
}
\keyword{datasets}
