% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MTCpairs2.R
\docType{data}
\name{MTCpairs2}
\alias{MTCpairs2}
\title{The MTCpairs2 data set}
\format{A data frame}
\usage{
data(MTCpairs2)
}
\description{
\itemize{
\item{Update}{ something}
\item{TE}{something}
\item{seTE}{something }
\item{studlab}{something }
\item{trea1}{ something}
\item{trea2}{ something}
\item{event1}{ something}
\item{n1}{something}
\item{event2}{ something}
\item{n2}{ something}
\item{id}{ something}
\item{vi}{ something}
\item{trt.pair}{tsomething}
}
}
\keyword{datasets}

