% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upreportdashboard.R
\name{upreportdashoard}
\alias{upreportdashoard}
\title{Meta-analysis reportshiny app, using shinydashboard (improved version)}
\usage{
upreportdashoard(initialprotocol = TRUE, initialreport = TRUE, pair=FALSE,
net = FALSE, data = NULL, outputformat = "pdf",clearproto = FALSE)
}
\arguments{
\item{initialprotocol}{logical value to indicate if it is the initial protocol, by default is TRUE.}

\item{initialreport}{logical value to indicate if it is the initial review, by default is TRUE.}

\item{pair}{logical value to indicate if pairwaise analysis should be run, by default it is FALSE. To run pair is needed to specify the data.}

\item{net}{logical value to indicate if the analysisi will include a network meta-analysis, by default it is FALSE.}

\item{data}{list with two components, a data frame with treatment information for the pairwise meta-analysis (treat1 and treat2), id to identify each observation
and trt.pair with the string name for the pairwise comparison in alphabetic order, generated using pairwise_metafor in data folder. The second
element is a list with the pairwise meta-analysis models generated using pairwise_metafor in data folder}

\item{outputformat}{format to download protocol and report}

\item{clearproto}{only keep the latest protocol file}
}
\description{
Meta-analysis reportshiny app, using shinydashboard (improved version)
}
\examples{
\dontrun{

upreportdashoard(initialprotocol = TRUE, initialreport = TRUE,pair =FALSE,
net = FALSE, data = NULL,outputformat="pdf", clearproto =FALSE)  

 upreportdashoard(initialprotocol = TRUE, initialreport = TRUE, pair =TRUE,
  net = FALSE, data = modstrMTCdata,outputformat = "pdf")
  
}
}
