% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaupdate.R
\name{metaupdate}
\alias{metaupdate}
\title{Meta-analysis shiny app}
\usage{
metaupdate(datapair, pair_result,trt.pair, treat1, treat2, id)
}
\arguments{
\item{datapair}{Data frame with treatment information for the pairwise meta-analysis (treat1 and treat2), id to identify each observation
and trt.pair with the string name for the pairwise comparison in alphabetic order, generated using pairwise_metafor in data folder}

\item{pair_result}{list with the pairwise meta-analysis models generated using pairwise_metafor in data folder}

\item{trt.pair}{variable name with the pairwise treatment names}

\item{treat1}{variable name with the treatment 1 in datapair}

\item{treat2}{variable name with the treatment 2 in datapair}

\item{id}{variable with id information in datapair}
}
\value{
shiny app.
}
\description{
\code{metaupdate} implements .....
}

