% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_metafor.R
\name{pairwise_metafor}
\alias{pairwise_metafor}
\title{Restruscture data and pairwise meta-analysis model results using metafor package}
\usage{
pairwise_metafor(dataini, ...)
}
\arguments{
\item{dataini}{Data frame in an arm-based format (e.g. input format for WinBUGS)}

\item{...}{optional argument to functions, you can include any parameter to run rma function from metafor pkg}
}
\value{
returns to .Rdata one with tde data set in contrast-based format and the second is a list with the pairwise meta analysis for each update and each pair of treatments
}
\description{
\code{pairwise_metafor} implements .....
}
\examples{
load("./data/MTCdata.Rdata")
MTCpairs <- pairwise(list(treat1, treat2, treat3),
                list(event1, event2, event3),
                list(n1, n2, n3),
                data = MTCdata,
                sm = "RR")
MTCpairs <- data.frame(up = c(rep(1, 109), rep(2, 5)), MTCpairs)
pairwise_metafor(MTCpairs,  method  = "REML")
}

