% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upreportdashboard.R
\name{upreportdashoard}
\alias{upreportdashoard}
\title{Meta-analysis reportshiny app, using shinydashboard (improved version)}
\usage{
upreportdashoard(initialprotocol = TRUE, initialreport = TRUE, pair=FALSE,
net = FALSE, data = NULL)
}
\arguments{
\item{initialprotocol}{logic value to indicate if is the initial protocol, default is TRUE}

\item{initialreport}{logic value to indicate if is the initial review, default is TRUE}

\item{pair}{logical, if FALSE implies regular analysis not pairwise}

\item{net}{logic value to indicate if the analysisi will include a network meta-analysis, default is FALSE}

\item{data}{list with two components, a data frame with treatment information for the pairwise meta-analysis (treat1 and treat2), id to identify each observation
and trt.pair with the string name for the pairwise comparison in alphabetic order, generated using pairwise_metafor in data folder. The second
element is a list with the pairwise meta-analysis models generated using pairwise_metafor in data folder}
}
\description{
Meta-analysis reportshiny app, using shinydashboard (improved version)
}
\examples{
\dontrun{

upreportdashoard(initialprotocol = TRUE, initialreport = TRUE,pair =FALSE,
net = FALSE, data = NULL)  
 upreportdashoard(initialprotocol = TRUE, initialreport = TRUE,pair =TRUE,
  net = FALSE, data = modstr)
}
}
